refactor: extracted location input component
